Initial Release
Quick-look maps of ammonia abundance, cloud pressure, and visual context are provided for the 2022, 2023-24, and 2024-25 apparitions of Jupiter using the band-average average method developed by Hill et al., 2024. The maps are composed of multiple observations spanning no more than a couple of nights of observing, and in 2024-25, often just a single night.

The files were created using continousmapscript.py.